Fix a bunch of incorrect code.
You know how you have something, and it's just purring along for
months, and then it starts to *die horribly*? And you look at the
code and exclaim "*WHAT* was I thinking?!". That's this update.

From trying to print a non-existent variable, to expecting
sys.print_exception to return a string, to printing red text on
a red background. Yeesh. Oh, and taking an exception and then
assuming that the function set variables.

Finally, just for fun, `wifi_read_from_flash()` no longer works in later
versions of micropython. It turns out, however, that `get_wifi()` works
from 1.5 on, so I've switched over to that.